Define protocols for serde for use in serialize decorator (step 1)
