+++ src/rootfilespec/bootstrap/TUUID.py
@@ -1,6 +1,7 @@
 from uuid import UUID
 
-from rootfilespec.structutil import ReadBuffer, ROOTSerializable, serializable
+from rootfilespec.serializable import serializable
+from rootfilespec.structutil import ReadBuffer, ROOTSerializable
 
 
 @serializable
