+++ src/rootfilespec/structutil.py
@@ -1,35 +1,12 @@
 import dataclasses
 import struct
-import sys
-from functools import partial
 from typing import (
-    Annotated,
     Any,
     Callable,
     Generic,
     Optional,
     TypeVar,
-    get_args,
-    get_origin,
-    get_type_hints,
 )
-
-import numpy as np
-from typing_extensions import dataclass_transform  # in typing for Python 3.11+
-
-
-def _get_annotations(cls: type) -> dict[str, Any]:
-    """Get the annotations of a class, including private attributes."""
-    if sys.version_info >= (3, 10):
-        from inspect import get_annotations
-
-        return get_annotations(cls)
-    return {
-        field: ann
-        for field, ann in cls.__dict__.get("__annotations__", {}).items()
-        if not field.startswith("_") and field != "self"
-    }
-
 
 Args = tuple[Any, ...]
 
@@ -130,80 +107,70 @@ class ReadBuffer:
 
 
 DataFetcher = Callable[[int, int], ReadBuffer]
-Members = dict[str, Any]
-ReadMethod = Callable[[ReadBuffer, Args], tuple[Args, ReadBuffer]]
+# Members = dict[str, Any]
 OutType = TypeVar("OutType")
+ReadMethod = Callable[[ReadBuffer, Args], tuple[Args, ReadBuffer]]
+ReadObjMethod = Callable[[ReadBuffer], tuple[OutType, ReadBuffer]]
 
 
-def _read_wrapper(cls: type["ROOTSerializable"]) -> ReadMethod:
-    """A wrapper to call the read method of a ROOTSerializable class."""
+class MemberSerDe:
+    """A protocol for Serialization/Deserialization method annotations for a field.
 
-    def read(buffer: ReadBuffer, args: Args) -> tuple[Args, ReadBuffer]:
-        obj, buffer = cls.read(buffer)
-        return (*args, obj), buffer
+    The @serializable decorator will use these annotations to determine how to read
+    the field from the buffer. For example, if a dataclass has a field of type
+        `field: Annotated[Type, MemberSerDe(*args)]`
+    Then `MemberSerDe.build_reader(Type)` will be called to get a function that
+    can read the field from the buffer.
+    """
 
-    return read
+    def build_reader(self, ftype: type) -> ReadMethod:
+        """Build a reader function for the given field type and arguments.
+
+        The reader function should take a ReadBuffer and return a tuple of the new
+        arguments and the remaining buffer.
+        """
+        msg = f"Cannot build reader for {self.__class__.__name__}"
+        raise NotImplementedError(msg)
+
+
+class ContainerSerDe(Generic[OutType]):
+    """A protocol for (De)serialization of generic container fields.
+
+    The @serializable decorator will use these annotations to determine how to read
+    the field from the buffer. For example, if a dataclass has a field of type
+        `field: Container[Type]`
+    Then `ContainerSerDe.build_reader(build_reader(Type))` will be called to get a function that
+    can read the field from the buffer.
+    """
+
+    def build_reader(self, inner_reader: ReadObjMethod[OutType]) -> ReadMethod:
+        """Build a reader function for the given field type and arguments.
+
+        The reader function should take a ReadBuffer and return a tuple of the new
+        arguments and the remaining buffer.
+        """
+        msg = f"Cannot build reader for {self.__class__.__name__}"
+        raise NotImplementedError(msg)
 
 
 @dataclasses.dataclass
-class Fmt:
+class _FmtReader:
+    fmt: str
+    outtype: type
+
+    def __call__(self, buffer: ReadBuffer, args: Args) -> tuple[Args, ReadBuffer]:
+        tup, buffer = buffer.unpack(self.fmt)
+        return (*args, self.outtype(*tup)), buffer
+
+
+@dataclasses.dataclass
+class Fmt(MemberSerDe):
     """A class to hold the format of a field."""
 
     fmt: str
 
-    def read_as(
-        self, outtype: type[OutType], buffer: ReadBuffer, args: Args
-    ) -> tuple[Args, ReadBuffer]:
-        tup, buffer = buffer.unpack(self.fmt)
-        return (*args, outtype(*tup)), buffer
-
-
-@dataclasses.dataclass
-class BasicArray:
-    """A class to hold a basic array of a given type.
-
-    Attributes:
-        dtype (np.dtype): The format of the array.
-        shapefield (str): The field in the parent object holding the
-            shape of the array.
-    """
-
-    dtype: np.dtype[Any]
-    shapefield: str
-
-
-@dataclasses.dataclass
-class _BasicArrayReadMethod:
-    dtype: np.dtype[Any]
-    sizeidx: int
-
-    def read(self, buffer: ReadBuffer, args: Args) -> tuple[Args, ReadBuffer]:
-        n = args[self.sizeidx]
-        pad, buffer = buffer.consume(1)
-        if not ((n == 0 and pad == b"\x00") or (n > 0 and pad == b"\x01")):
-            msg = f"Expected null or 0x01 pad byte but got {pad!r} for size {n}"
-            raise ValueError(msg)
-        data, buffer = buffer.consume(n * self.dtype.itemsize)
-        arg = np.frombuffer(data, dtype=self.dtype, count=n)
-        return (*args, arg), buffer
-
-
-@dataclasses.dataclass
-class FixedSizeArray:
-    """A class to hold a fixed size array of a given type.
-
-    Attributes:
-        dtype (np.dtype): The format of the array.
-        size (int): The size of the array.
-    """
-
-    dtype: np.dtype[Any]
-    size: int
-
-    def read(self, buffer: ReadBuffer, args: Args) -> tuple[Args, ReadBuffer]:
-        data, buffer = buffer.consume(self.size * self.dtype.itemsize)
-        arg = np.frombuffer(data, dtype=self.dtype, count=self.size)
-        return (*args, arg), buffer
+    def build_reader(self, ftype: type) -> ReadMethod:
+        return _FmtReader(self.fmt, ftype)
 
 
 T = TypeVar("T", bound="ROOTSerializable")
@@ -224,133 +191,3 @@ class ROOTSerializable:
     def read_members(cls, buffer: ReadBuffer) -> tuple[Args, ReadBuffer]:
         msg = "Unimplemented method: {cls.__name__}.read_members"
         raise NotImplementedError(msg)
-
-
-@dataclasses.dataclass
-class Pointer(ROOTSerializable, Generic[T]):
-    obj: Optional[T]
-
-    @classmethod
-    def read(cls, buffer: ReadBuffer):
-        (addr,), buffer = buffer.unpack(">i")
-        if not addr:
-            return cls(None), buffer
-        # TODO: use read_streamed_item to read the object
-        if addr & 0x40000000:
-            # this isn't actually an address but an object
-            addr &= ~0x40000000
-            # skip forward
-            buffer = buffer[addr:]
-        return cls(None), buffer
-
-
-@dataclasses.dataclass
-class StdVector(ROOTSerializable, Generic[T]):
-    """A class to represent a std::vector<T>.
-
-    Attributes:
-        items (list[T]): The list of objects in the vector.
-    """
-
-    items: tuple[T, ...]
-
-    @classmethod
-    def read_as(
-        cls, outtype: type[T], buffer: ReadBuffer, args: Args
-    ) -> tuple[Args, ReadBuffer]:
-        (n,), buffer = buffer.unpack(">i")
-        out: tuple[T, ...] = ()
-        if outtype is StdVector:
-            (interior_type,) = get_args(outtype)
-            for _ in range(n):
-                out, buffer = StdVector.read_as(interior_type, buffer, out)
-        elif getattr(outtype, "_name", None) == "Annotated":
-            # TODO: this should be handled in the serializable decorator
-            (ftype, fmt) = get_args(outtype)
-            if isinstance(fmt, Fmt):
-                for _ in range(n):
-                    out, buffer = fmt.read_as(ftype, buffer, out)
-            else:
-                msg = f"Cannot read field of type {outtype} with format {fmt}"
-                raise NotImplementedError(msg)
-        else:
-            for _ in range(n):
-                obj, buffer = outtype.read(buffer)
-                out += (obj,)
-        return (*args, cls(out)), buffer
-
-
-@dataclass_transform()
-def serializable(cls: type[T]) -> type[T]:
-    """A decorator to add a read_members method to a class that reads its fields from a buffer.
-
-    The class must have type hints for its fields, and the fields must be of types that
-    either have a read method or are subscripted with a Fmt object.
-    """
-    cls = dataclasses.dataclass(eq=False)(cls)
-
-    # if the class already has a read_members method, don't overwrite it
-    readmethod = getattr(cls, "read_members", None)
-    if (
-        readmethod
-        and getattr(readmethod, "__qualname__", None)
-        == f"{cls.__qualname__}.read_members"
-    ):
-        return cls
-
-    names: list[str] = []
-    constructors: list[ReadMethod] = []
-    namespace = get_type_hints(cls, include_extras=True)
-    for field in _get_annotations(cls):
-        names.append(field)
-        ftype = namespace[field]
-        if isinstance(ftype, type) and issubclass(ftype, ROOTSerializable):
-            constructors.append(_read_wrapper(ftype))
-        elif origin := get_origin(ftype):
-            if origin is Annotated:
-                ftype, *annotations = get_args(ftype)
-                if not annotations:
-                    msg = f"Cannot read field {field} of type {ftype} (missing annotations)"
-                    raise NotImplementedError(msg)
-                format, *annotations = annotations
-                if isinstance(format, Fmt):
-                    # TODO: potential optimization: consecutive struct fields could be read in one struct.unpack call
-                    constructors.append(partial(format.read_as, ftype))
-                elif isinstance(format, BasicArray):
-                    assert ftype is np.ndarray
-                    if format.shapefield not in names:
-                        # TODO: to implement this we need to migrate read_members to return Members rather than Args
-                        msg = f"Cannot yet read {field} because shape field {format.shapefield} is in base class"
-                        raise NotImplementedError(msg)
-                    fieldindex = names.index(format.shapefield)
-                    constructors.append(
-                        _BasicArrayReadMethod(format.dtype, fieldindex).read
-                    )
-                elif isinstance(format, FixedSizeArray):
-                    assert ftype is np.ndarray
-                    constructors.append(format.read)
-                else:
-                    msg = f"Cannot read field {field} of type {ftype} with format {format}"
-                    raise NotImplementedError(msg)
-            elif origin is Pointer:
-                constructors.append(_read_wrapper(origin))
-            elif origin is StdVector:
-                (ftype,) = get_args(ftype)
-                # TODO: nested std::vectors here instead of in StdVector.read_as
-                constructors.append(partial(StdVector.read_as, ftype))
-            else:
-                msg = f"Cannot read field {field} of subscripted type {ftype} with origin {origin}"
-                raise NotImplementedError(msg)
-        else:
-            msg = f"Cannot read field {field} of type {ftype}"
-            raise NotImplementedError(msg)
-
-    @classmethod  # type: ignore[misc]
-    def read_members(_: type[T], buffer: ReadBuffer) -> tuple[Args, ReadBuffer]:
-        args: Args = ()
-        for constructor in constructors:
-            args, buffer = constructor(buffer, args)
-        return args, buffer
-
-    cls.read_members = read_members  # type: ignore[assignment]
-    return cls
